Distrobuilder configuration
I use Distrobuilder to build a minimal debootstrap(1) image that is then
added to the LXD image store. Yes I could use debootstrap itself but
this means I can be lazy!

This probably isn't any use to anyone other than rgh because it uses my
local aptcacher repo, though it wouldn't be hard to change it or remove
it altogether (which will make the whole process *much* slower).