Basic terraform file
This creates three containes: postgres, rails and wap and serves as a
proof of concept to see how we go with Terraform.

It is needlessly long as there are stanzas that are there simply because
of my setup (no DNS, for example) plus I can reduce the repetition of
each stanza with modules.