Attempt to implement modules
This doesn't work. It's not possible to parameterise resources so it's
simply not possible to do what I've tried to do. I assume there is
another mechanism to this by using either dynamic blocks—which I don't
quite understand—or templates.

So although some of the details will change this is roughly how the
terraform files should be.