Reorganise how the scripts are run
Having all the scripts run from one "scripts" stanza is a bit of nightmare
especially given that there are dependencies that shouldn't be exposed
to the end user.

This commit splits the scripts that should be run first and last into
separate stanzas so it is clearer. These can also be hidden from the end
user.