Use a more reliable apt mirror
This is the main devuan mirror so it might be slow at times but given we
are in the arse end of nowhere I suspect we'll be right. Plus it does
actually do what it says on the box which is more than can be said for
some of the mirrors!